web/feat: move isLoading animation to page
