+++ web/app/components/Sigma/SigmaAccordion.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { useSigmaYesterday } from '@/app/lib/api/clientApi'
 import * as Accordion from '@radix-ui/react-accordion'
 import clsx from 'clsx'
 import { AnimatePresence } from 'framer-motion'
@@ -32,26 +31,6 @@ export function SigmaAccordion({}: {}) {
 		/>
 	))
 
-	const { isLoading } = useSigmaYesterday({
-		marketCap,
-		page: pageIndex,
-		typeLabels,
-	})
-
-	const [loadingAnimation, setLoadingAnimation] = useState<boolean>(false)
-
-	const handleAnimationIteration = () => {
-		if (!isLoading) {
-			setLoadingAnimation(false)
-		}
-	}
-
-	useEffect(() => {
-		if (isLoading) {
-			setLoadingAnimation(true)
-		}
-	}, [isLoading])
-
 	return (
 		<div className="grid-cols-default sm:grid">
 			<div
@@ -59,9 +38,7 @@ export function SigmaAccordion({}: {}) {
 					'col-start-2 -mx-2 box-border h-[50rem] overflow-x-hidden overflow-y-scroll px-2 py-2 transition-all duration-1000 sm:rounded-2xl',
 					pageIndex > 1 &&
 						'bg-gradient-to-b from-slate-a2 via-transparent to-slate-a2',
-					loadingAnimation && 'animate-pulse',
 				)}
-				onAnimationIteration={handleAnimationIteration}
 			>
 				<Accordion.Root
 					collapsible
